Update and rename Sprites.py to sprites.py
